dernier objet acheté, vue pour la création d'objets, retrait d'objets de la vente

+++ tp4/sellyours/controllers/buyController.js
@@ -12,15 +12,20 @@ const buyItem = async (req, res) => {
     const seller = await userSchema.findById(sellerId)
 
     const price = item.price
-    const newBuyerBalance = { balance : buyer.balance - price }
-    const newSellerBalance = { balance : seller.balance + price }
+    if(buyer.balance > price) {
+        const newBuyerBalance = { balance : buyer.balance - price }
+        const newSellerBalance = { balance : seller.balance + price }
 
-    const updatedBuyer = await userSchema.findByIdAndUpdate(buyerId, newBuyerBalance, { new : true })
-    const updatedSeller = await userSchema.findByIdAndUpdate(sellerId, newSellerBalance, { new : true })
+        const updatedBuyer = await userSchema.findByIdAndUpdate(buyerId, newBuyerBalance, { new : true })
+        const updatedSeller = await userSchema.findByIdAndUpdate(sellerId, newSellerBalance, { new : true })
 
-    const deleteItem = await itemSchema.findByIdAndDelete(itemId)
+        const deleteItem = await itemSchema.findByIdAndDelete(itemId)
 
-    res.status(200).json(updatedBuyer)
+        res.status(200).json(item)
+    }
+    else {
+        res.status(401).json( { message : "Vous n'avez pas assez d'argent pour acheter cet objet."} )
+    }
 }
 
 module.exports.buyItem = buyItem
